Remove dg.commands.ibmcloud.common module
