+++ dg/commands/ibmcloud/cluster/create.py
@@ -14,20 +14,20 @@
 
 import click
 
-from dg.commands.ibmcloud.common import (
-    get_default_private_vlan,
-    get_default_public_vlan,
-    is_logged_in,
-)
 from dg.lib.ibmcloud import execute_ibmcloud_command_without_check
 from dg.lib.ibmcloud.cluster.rm import delete_ibmcloud_cluster
 from dg.lib.ibmcloud.install import install_cp4d_with_preinstall
+from dg.lib.ibmcloud.login import is_logged_in
 from dg.lib.ibmcloud.login import login as login_to_ibm_cloud
 from dg.lib.ibmcloud.oc import get_latest_supported_openshift_version
 from dg.lib.ibmcloud.status import (
     cluster_exists,
     wait_for_cluster_deletion,
     wait_for_cluster_readiness,
+)
+from dg.lib.ibmcloud.vlan import (
+    get_default_private_vlan,
+    get_default_public_vlan,
 )
